# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "reeve",
    "description": "Fleet desired-state manager (Margo-inspired). Human API consumed by the reeve UI, plus the device-facing wire surface.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/join-tokens": {
      "get": {
        "tags": [
          "join-tokens"
        ],
        "summary": "GET /api/join-tokens (admin/operator) — hashes and metadata only.",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "All join tokens, newest first (hashes and metadata only)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/JoinTokenInfo"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below operator role"
          }
        }
      },
      "post": {
        "tags": [
          "join-tokens"
        ],
        "summary": "POST /api/join-tokens (admin/operator, D4). Returns the raw token —\nshown exactly once.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateJoinTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Join token created; raw token shown exactly once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreatedJoinToken"
                }
              }
            }
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below operator role"
          },
          "404": {
            "description": "Unknown device_id for a re-enroll token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Non-positive ttl_secs or max_uses",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
    "/api/join-tokens/{token_hash}": {
      "delete": {
        "tags": [
          "join-tokens"
        ],
        "summary": "DELETE /api/join-tokens/{token_hash} (admin/operator) — revoke.\nIdempotent: unknown or already-revoked is still 204.",
        "operationId": "delete",
        "parameters": [
          {
            "name": "token_hash",
            "in": "path",
            "description": "Hex sha256 of the join token",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Revoked (idempotent: unknown or already-revoked is still 204)"
          },
          "401": {
            "description": "Unauthenticated"
          },
          "403": {
            "description": "Below operator role"
          }
        }
      }
    },
...
    "/api/reeve/v1/enroll": {
      "post": {
        "tags": [
          "enroll"
        ],
        "summary": "POST /api/reeve/v1/enroll (D4 step 1-2).",
        "operationId": "enroll_route",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EnrollRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Enrolled (or idempotently resumed); the device credential is shown exactly once",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnrollResponse"
                }
              }
            }
          },
          "401": {
            "description": "Invalid, expired, or exhausted join token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "422": {
            "description": "Malformed enrollment request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          },
          "500": {
            "description": "Enrollment failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateJoinTokenRequest": {
        "type": "object",
        "properties": {
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Existing device to bind a re-enroll token to (D4)."
          },
          "max_uses": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Default 1 (D4)."
          },
          "ttl_secs": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64",
            "description": "Seconds until expiry; default 86400 (24h, D4)."
          }
        }
      },
...
      "CreatedJoinToken": {
        "type": "object",
        "description": "`POST /api/join-tokens` body: the raw token, shown exactly once.",
        "required": [
          "join_token",
          "token_hash",
          "expires_at",
          "max_uses"
        ],
        "properties": {
          "device_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "expires_at": {
            "type": "integer",
            "format": "int64"
          },
          "join_token": {
            "type": "string",
            "description": "The raw join token (`rvj_<64 hex>`) — only the hash is stored."
          },
          "max_uses": {
            "type": "integer",
            "format": "int64"
          },
          "token_hash": {
            "type": "string"
          }
        }
      },
...
      "EnrollRequest": {
        "type": "object",
        "description": "Request body of `POST /api/reeve/v1/enroll` (D4 step 1):\n`{ join_token, hostname, arch, agent_version }`.",
        "required": [
          "join_token",
          "hostname",
          "arch",
          "agent_version"
        ],
        "properties": {
          "agent_version": {
            "type": "string",
            "description": "reeve-agent version performing the enrollment."
          },
          "arch": {
            "type": "string",
            "description": "Target architecture (e.g. `x86_64`, `aarch64`)."
          },
          "hostname": {
            "type": "string",
            "description": "The enrolling box's hostname — the idempotency key for a retried\ninstall with the same join token (D4)."
          },
          "join_token": {
            "type": "string",
            "description": "Operator-created join token (D4: TTL + max-uses, stored hashed\nserver-side). A re-enroll token is the same grammar, bound to an\nexisting device_id server-side."
          }
        }
      },
      "EnrollResponse": {
        "type": "object",
        "description": "Response of `POST /api/reeve/v1/enroll` (D4 step 2): the issued\nidentity and the ONE device credential for every device-facing\nsurface (docs/decisions/auth.md D1).",
        "required": [
          "device_id",
          "device_token"
        ],
        "properties": {
          "device_id": {
            "type": "string",
            "description": "Server-issued device identity."
          },
          "device_token": {
            "type": "string",
            "description": "Enrollment-issued bearer token (`rvd_<64 hex>`). Shown exactly\nonce; the server keeps only its hash."
          },
          "resumed": {
            "type": "boolean",
            "description": "True when this enrollment resumed an existing device identity\n(idempotent re-run of the same join token + hostname, or a\nre-enroll token bound to an existing device — D4)."
          }
        }
      },
      "ErrorBody": {
        "type": "object",
        "description": "The one error-body shape every JSON error response uses:\n`{ \"error\": \"<human-readable message>\" }`. Schema component for the\nD10 openapi.json (docs/decisions/ui.md); handlers keep building the\nbody via `json!` — this type documents the contract.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "JoinTokenInfo": {
        "type": "object",
        "description": "One join token row, as listed to operators (never the raw token).",
        "required": [
          "token_hash",
          "created_by",
          "created_at",
          "expires_at",
          "max_uses",
          "uses"
        ],
        "properties": {
          "created_at": {
            "type": "integer",
            "format": "int64"
          },
          "created_by": {
            "type": "string"
          },
          "device_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Set => re-enroll token bound to this existing device (D4)."
          },
          "expires_at": {
            "type": "integer",
            "format": "int64"
          },
          "max_uses": {
            "type": "integer",
            "format": "int64"
          },
          "revoked_at": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          },
          "token_hash": {
            "type": "string"
          },
          "uses": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    }
  },
...
}